+++ components/analytics/shake2report/impl/src/main/java/com/flipperdevices/analytics/shake2report/impl/model/Shake2ReportState.kt
@@ -1,0 +1,8 @@
+package com.flipperdevices.analytics.shake2report.impl.model
+
+sealed class Shake2ReportState {
+    object Pending : Shake2ReportState()
+    object Compressing : Shake2ReportState()
+    object Uploading : Shake2ReportState()
+    data class Complete(val id: String) : Shake2ReportState()
+}