+++ components/analytics/shake2report/impl/src/main/java/com/flipperdevices/analytics/shake2report/impl/viewmodel/Shake2ReportViewModel.kt
@@ -1,0 +1,112 @@
+package com.flipperdevices.analytics.shake2report.impl.viewmodel
+
+import android.app.Application
+import android.content.ClipData
+import android.content.ClipboardManager
+import androidx.core.content.ContextCompat
+import androidx.lifecycle.AndroidViewModel
+import androidx.lifecycle.viewModelScope
+import com.flipperdevices.analytics.shake2report.impl.InternalShake2Report
+import com.flipperdevices.analytics.shake2report.impl.R
+import com.flipperdevices.analytics.shake2report.impl.model.Shake2ReportState
+import com.flipperdevices.core.ktx.android.toast
+import io.sentry.Attachment
+import io.sentry.Sentry
+import io.sentry.SentryEvent
+import io.sentry.SentryLevel
+import io.sentry.protocol.Message
+import io.sentry.protocol.SentryId
+import kotlinx.coroutines.Dispatchers
+import kotlinx.coroutines.flow.MutableStateFlow
+import kotlinx.coroutines.flow.asStateFlow
+import kotlinx.coroutines.launch
+import kotlinx.coroutines.withContext
+import net.lingala.zip4j.ZipFile
+import net.lingala.zip4j.model.ZipParameters
+import net.lingala.zip4j.model.enums.CompressionLevel
+import net.lingala.zip4j.model.enums.CompressionMethod
+import tangle.viewmodel.VMInject
+import java.io.File
+import java.util.concurrent.TimeUnit
+
+@Suppress("MagicNumber")
+private val SENTRY_TIMEOUT_MS = TimeUnit.MILLISECONDS.convert(5, TimeUnit.MINUTES)
+
+class Shake2ReportViewModel @VMInject constructor(
+    application: Application,
+    private val internalShake2Report: InternalShake2Report
+) : AndroidViewModel(application) {
+    private val shake2ReportStateFlow = MutableStateFlow<Shake2ReportState>(
+        Shake2ReportState.Pending
+    )
+
+    fun getState() = shake2ReportStateFlow.asStateFlow()
+
+    fun report(
+        name: String,
+        description: String
+    ) = viewModelScope.launch(Dispatchers.Default) {
+        if (shake2ReportStateFlow.value !is Shake2ReportState.Pending) {
+            return@launch
+        }
+        shake2ReportStateFlow.emit(Shake2ReportState.Compressing)
+        val compressingFile = compressLogFolder()
+        shake2ReportStateFlow.emit(Shake2ReportState.Uploading)
+        val sentryId = sendingReport(name, description, compressingFile)
+        shake2ReportStateFlow.emit(Shake2ReportState.Complete(sentryId))
+    }
+
+    fun copyToClipboard(id: String) {
+        val context = getApplication<Application>()
+        val clipboardManager = ContextCompat.getSystemService(context, ClipboardManager::class.java)
+        val clipData = ClipData.newHtmlText(
+            context.getString(R.string.shake2report_copy_sentry_title),
+            id,
+            "<code>$id</code>"
+        )
+        clipboardManager?.setPrimaryClip(clipData)
+        context.toast(R.string.shake2report_copy_sentry_copy_toast)
+    }
+
+    @Suppress("BlockingMethodInNonBlockingContext")
+    private suspend fun compressLogFolder(): File = withContext(Dispatchers.IO) {
+        val logDir = internalShake2Report.logDir
+        val zipFile = ZipFile(
+            File(
+                getApplication<Application>().cacheDir,
+                "logs-${System.currentTimeMillis()}.zip"
+            )
+        )
+        val zipParameters = ZipParameters().apply {
+            compressionMethod = CompressionMethod.DEFLATE
+            compressionLevel = CompressionLevel.ULTRA
+        }
+        zipFile.addFolder(logDir, zipParameters)
+
+        return@withContext zipFile.file
+    }
+
+    private suspend fun sendingReport(
+        name: String,
+        description: String,
+        logZip: File
+    ): String = withContext(Dispatchers.IO) {
+        val event = SentryEvent()
+        event.level = SentryLevel.ERROR
+        event.message = Message().apply {
+            message = name.ifBlank {
+                "Error via shake2report ${System.currentTimeMillis()}"
+            }
+            params = (params ?: emptyList()).plus(description.split("\n"))
+        }
+
+        lateinit var sentryId: SentryId
+        Sentry.withScope { scope ->
+            scope.addAttachment(Attachment(logZip.absolutePath))
+            sentryId = Sentry.captureEvent(event)
+            Sentry.flush(SENTRY_TIMEOUT_MS)
+            scope.clearAttachments()
+        }
+        return@withContext sentryId.toString()
+    }
+}